Added Rating component. Added an ability to set article rating for authorized users. Fixed error with showing duplicate comments. Fixed error with rating value mapping.

+++ BlogPlatform.Domain/Entities/Rating.cs
@@ -5,7 +5,7 @@ namespace BlogPlatform.Domain.Entities
     public class Rating : IEntityBase
     {
         public int Id { get; set; }
-        public byte Value { get; set; }
+        public double Value { get; set; }
         public DateTime DateAdded { get; set; }
         public int ArticleId { get; set; }
         public int AccountId { get; set; }
